address some of the review comments
